changed int results to double

+++ Calculator/src/com/qa/calculator/Result.java
@@ -13,7 +13,7 @@ public class Result {
 		
 	}
 	
-	public static int method1(int num1, int num2, int num3) {
+	public static double method1(int num1, int num2, int num3) {
 		biology = num3;
 		chemistry = num2;
 		physics = num1;
@@ -21,7 +21,7 @@ public class Result {
 		int exam1 = physics;
 		int exam2 = chemistry;
 		int exam3 = biology;
-		int result = (int) total;
+		double result = total;
 		
 		System.out.println("Exam 1 = " + exam1 + ", Exam 2 = " + exam2 + ", Exam 3 = " + exam3);
 		System.out.println("Total Score = " + result);
